xhs-crawler/requirements.txt
# 环境变量
python-dotenv>=1.0.0

# 浏览器自动化（阶段 1）
playwright>=1.40.0

# MySQL 写入（阶段 3）
PyMySQL>=1.0.0
